refactor: folder structure

+++ app/admin/manage-clinics/page.tsx
@@ -13,11 +13,13 @@ import { LuClipboardEdit } from "react-icons/lu";
 import Cookies from "js-cookie";
 import toast from "react-hot-toast";
 
+import { deleteClinic, getAllClinics } from "@/services/clinic-service";
+
+import { Province } from "@/types/auth-types";
+import { ClinicData } from "@/types/clinic-types";
+
 import useDebounce from "@/hooks/use-debounce";
 import useProvinces from "@/hooks/fetch/use-provinces";
-
-import { ClinicData } from "@/types/clinic-types";
-import { deleteClinic, getAllClinics } from "@/services/clinic-service";
 
 import {
   Table,
@@ -38,7 +40,7 @@ import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
 import Spinner from "@/components/spinner";
 import PaginationSection from "@/components/pagination";
-import DeleteConfirmationModal from "@/components/delete-confirmation-modal";
+import DeleteConfirmationModal from "@/components/modal/delete-confirmation-modal";
 
 const ManageClinics = () => {
   const router = useRouter();
@@ -47,7 +49,7 @@ const ManageClinics = () => {
   const [inputFocused, setInputFocused] = useState<boolean>(false);
   const [isLoading, setLoading] = useState({ clinics: false, deleting: false });
 
-  const provinces: any[] = useProvinces();
+  const provinces: Province[] = useProvinces();
   const [selectedProvince, setSelectedProvince] = useState<string>("all");
 
   const [clinics, setClinics] = useState<ClinicData[]>([]);
@@ -62,24 +64,26 @@ const ManageClinics = () => {
   const fetchClinics = async (page: number, query?: string, province?: string) => {
     const accessToken = Cookies.get("access_token");
     if (!accessToken) return;
-
     setLoading({ ...isLoading, clinics: true });
 
     try {
       const { clinics, total } = await getAllClinics({
         page, limit: 10, query, exclude: "desc, banner", province
       });
-
-      const itemsPerPage = 10;
-      const totalPages = Math.ceil(total / itemsPerPage);
-
-      setClinics(clinics);
-      setTotalPages(totalPages);
+      handleClinicsFetchSuccess(clinics, total);
     } catch (error: any) {
+      console.error(error);
       toast.error("Có lỗi xảy ra. Vui lòng thử lại sau ít phút nữa!");
     } finally {
       setLoading({ ...isLoading, clinics: false });
     }
+  };
+
+  const handleClinicsFetchSuccess = (clinics: ClinicData[], total: number) => {
+    const itemsPerPage = 10;
+    const totalPages = Math.ceil(total / itemsPerPage);
+    setClinics(clinics);
+    setTotalPages(totalPages);
   };
 
   useEffect(() => {
@@ -87,29 +91,30 @@ const ManageClinics = () => {
   }, [currentPage, debouncedSearchQuery, selectedProvince]);
 
   const handleDeleteClinic = async () => {
-    if (!clinicToDelete) return;
-
     const accessToken = Cookies.get("access_token");
-    if (!accessToken) return;
-
+    if (!accessToken || !clinicToDelete) return;
     setLoading({ ...isLoading, deleting: true });
 
     try {
       await deleteClinic(accessToken, clinicToDelete);
-
-      if (currentPage > 1 && clinics?.length === 1) {
-        setCurrentPage(currentPage - 1);
-        await fetchClinics(currentPage - 1);
-      } else {
-        await fetchClinics(currentPage);
-      }
-
+      await handlePrefetchingClinics();
       setModalOpen(false);
       toast.success("Xóa bệnh viện thành công!");
     } catch (error: any) {
+      console.error(error);
       toast.error("Xóa bệnh viện thất bại. Vui lòng thử lại sau ít phút nữa!");
     } finally {
+      setModalOpen(false);
       setLoading({ ...isLoading, deleting: false });
+    }
+  };
+
+  const handlePrefetchingClinics = async () => {
+    if (currentPage > 1 && clinics?.length === 1) {
+      setCurrentPage(currentPage - 1);
+      await fetchClinics(currentPage - 1);
+    } else {
+      await fetchClinics(currentPage);
     }
   };
 
@@ -120,25 +125,20 @@ const ManageClinics = () => {
   return (
     <>
       <h1 className="text-xl font-bold mb-4">Danh sách bệnh viện</h1>
-
       <div className="flex items-center justify-between">
         <Select onValueChange={(value) => setSelectedProvince(value)}>
           <SelectTrigger className="w-[220px] p-3 border-none shadow-none">
             <div className="flex items-center gap-3">
               <FaLocationDot className="size-5 text-[#1c3f66]" />
-              <SelectValue
-                placeholder={selectedProvince === "all" ? "Tất cả" : selectedProvince || "Chọn tỉnh thành"}
-              />
+              <SelectValue placeholder={selectedProvince === "all" ? "Tất cả" : selectedProvince || "Chọn tỉnh thành"} />
             </div>
           </SelectTrigger>
           <SelectContent>
             {provinces?.length > 0 ? (
               <>
                 <SelectItem value="all">Tất cả</SelectItem>
-                {provinces?.map((province) => (
-                  <SelectItem key={province?.id} value={province?.name}>
-                    {province?.name}
-                  </SelectItem>
+                {provinces?.map((province: any) => (
+                  <SelectItem key={province?.id} value={province?.name}>{province?.name}</SelectItem>
                 ))}
               </>
             ) : (
@@ -172,8 +172,7 @@ const ManageClinics = () => {
       </div>
 
       <div className={cn(
-        "relative rounded-md shadow-md overflow-x-auto",
-        clinics?.length > 0 ? "h-auto" : "h-full"
+        "relative rounded-md shadow-md overflow-x-auto", clinics?.length > 0 ? "h-auto" : "h-full"
       )}>
         <Table className="relative h-full text-[17px]">
           <TableHeader className="sticky top-0 left-0 right-0 h-12 bg-gray-100">
@@ -188,7 +187,7 @@ const ManageClinics = () => {
 
           <TableBody>
             {clinics?.length > 0 ? (
-              clinics?.map((clinic, index) => {
+              clinics?.map((clinic: ClinicData, index) => {
                 const startIndex = (currentPage - 1) * 10;
 
                 return (
@@ -210,15 +209,10 @@ const ManageClinics = () => {
                         <Button
                           type="button"
                           variant="ghost"
-                          onClick={() =>
-                            router.replace(`/admin/manage-clinics/edit-clinic/${btoa(clinic?._id)}`)
-                          }
+                          onClick={() => router.replace(`/admin/manage-clinics/edit-clinic/${btoa(clinic?._id)}`)}
                           className="group"
                         >
-                          <LuClipboardEdit
-                            size="22"
-                            className="group-hover:text-green-500 transition duration-500"
-                          />
+                          <LuClipboardEdit size="22" className="group-hover:text-green-500 transition duration-500" />
                         </Button>
 
                         <Button
@@ -230,10 +224,7 @@ const ManageClinics = () => {
                           }}
                           className="group"
                         >
-                          <AiOutlineDelete
-                            size="22"
-                            className="group-hover:text-red-500 transition duration-500"
-                          />
+                          <AiOutlineDelete size="22" className="group-hover:text-red-500 transition duration-500" />
                         </Button>
                       </div>
                     </TableCell>
