+++ app/system/manage-posts/page.tsx
@@ -32,7 +32,7 @@ import { DatePicker } from "@/components/ui/date-picker";
 import Hint from "@/components/hint";
 import Spinner from "@/components/spinner";
 import PaginationSection from "@/components/pagination";
-import DeleteConfirmationModal from "@/components/delete-confirmation-modal";
+import DeleteConfirmationModal from "@/components/modal/delete-confirmation-modal";
 
 const ManagePosts = () => {
   const router = useRouter();
@@ -57,7 +57,6 @@ const ManagePosts = () => {
 
     const decodedToken: any = jwtDecode<JwtPayload>(accessToken);
     if (!decodedToken) return;
-
     setLoading({ ...isLoading, posts: true });
 
     try {
@@ -69,17 +68,20 @@ const ManagePosts = () => {
         exclude: "desc, image",
         releaseDate: date && format(date, "dd/MM/yyyy")
       });
-
-      const itemsPerPage = 10;
-      const totalPages = Math.ceil(total / itemsPerPage);
-
-      setPosts(posts);
-      setTotalPages(totalPages);
+      handlePostsFetchSuccess(posts, total);
     } catch (error: any) {
+      console.error(error);
       toast.error("Có lỗi xảy ra. Vui lòng thử lại sau ít phút nữa!");
     } finally {
       setLoading({ ...isLoading, posts: false });
     }
+  };
+
+  const handlePostsFetchSuccess = (posts: PostData[], total: number) => {
+    const itemsPerPage = 10;
+    const totalPages = Math.ceil(total / itemsPerPage);
+    setPosts(posts);
+    setTotalPages(totalPages);
   };
 
   useEffect(() => {
@@ -87,29 +89,30 @@ const ManagePosts = () => {
   }, [currentPage, debouncedSearchQuery, selectedDate]);
 
   const handleDeletePost = async () => {
-    if (!postToDelete) return;
-
     const accessToken = Cookies.get("access_token");
-    if (!accessToken) return;
-
+    if (!accessToken || !postToDelete) return;
     setLoading({ ...isLoading, deleting: true });
 
     try {
       await deletePost(accessToken, postToDelete);
-
-      if (currentPage > 1 && posts?.length === 1) {
-        setCurrentPage(currentPage - 1);
-        await fetchPosts(currentPage - 1);
-      } else {
-        await fetchPosts(currentPage);
-      }
-
+      await handlePrefetchingPosts();
       setModalOpen(false);
       toast.success("Xóa bài viết thành công!");
     } catch (error: any) {
+      console.error(error);
       toast.error("Xóa bài viết thất bại. Vui lòng thử lại sau ít phút nữa!");
     } finally {
+      setModalOpen(false);
       setLoading({ ...isLoading, deleting: false });
+    }
+  };
+
+  const handlePrefetchingPosts = async () => {
+    if (currentPage > 1 && posts?.length === 1) {
+      setCurrentPage(currentPage - 1);
+      await fetchPosts(currentPage - 1);
+    } else {
+      await fetchPosts(currentPage);
     }
   };
 
@@ -120,7 +123,6 @@ const ManagePosts = () => {
   return (
     <>
       <h1 className="text-xl font-bold mb-4">Danh sách bài viết</h1>
-
       <div className="flex items-center justify-between">
         <DatePicker
           selectedDate={selectedDate}
@@ -152,8 +154,7 @@ const ManagePosts = () => {
       </div>
 
       <div className={cn(
-        "relative rounded-md shadow-md overflow-x-auto",
-        posts?.length > 0 ? "h-auto" : "h-full"
+        "relative rounded-md shadow-md overflow-x-auto", posts?.length > 0 ? "h-auto" : "h-full"
       )}>
         <Table className="relative h-full text-[17px]">
           <TableHeader className="sticky top-0 left-0 right-0 h-12 bg-gray-100">
@@ -169,7 +170,7 @@ const ManagePosts = () => {
 
           <TableBody>
             {posts?.length > 0 ? (
-              posts?.map((post, index) => {
+              posts?.map((post: PostData, index) => {
                 const startIndex = (currentPage - 1) * 10;
 
                 return (
@@ -205,10 +206,7 @@ const ManagePosts = () => {
                           onClick={() => router.replace(`/system/manage-posts/edit-post/${btoa(post?._id)}`)}
                           className="group"
                         >
-                          <LuClipboardEdit
-                            size="22"
-                            className="group-hover:text-green-500 transition duration-500"
-                          />
+                          <LuClipboardEdit size="22" className="group-hover:text-green-500 transition duration-500" />
                         </Button>
 
                         <Button
@@ -220,10 +218,7 @@ const ManagePosts = () => {
                           }}
                           className="group"
                         >
-                          <AiOutlineDelete
-                            size="22"
-                            className="group-hover:text-red-500 transition duration-500"
-                          />
+                          <AiOutlineDelete size="22" className="group-hover:text-red-500 transition duration-500" />
                         </Button>
                       </div>
                     </TableCell>
